Implemented word document extraction
